+++ lib/src/analyzers/unused_code_analyzer/unused_code_analyzer.dart
@@ -13,6 +13,7 @@ import '../../config_builder/config_builder.dart';
 import '../../config_builder/models/analysis_options.dart';
 import '../../reporters/models/reporter.dart';
 import '../../utils/analyzer_utils.dart';
+import '../../utils/suppression.dart';
 import 'models/file_elements_usage.dart';
 import 'models/unused_code_file_report.dart';
 import 'models/unused_code_issue.dart';
@@ -25,6 +26,8 @@ import 'used_code_visitor.dart';
 
 /// The analyzer responsible for collecting unused code reports.
 class UnusedCodeAnalyzer {
+  static const _ignoreName = 'unused-code';
+
   const UnusedCodeAnalyzer();
 
   /// Returns a reporter for the given [name]. Use the reporter
@@ -126,7 +129,13 @@ class UnusedCodeAnalyzer {
 
   Set<Element> _analyzeFilePublicCode(SomeResolvedUnitResult unit) {
     if (unit is ResolvedUnitResult) {
-      final visitor = PublicCodeVisitor();
+      final suppression = Suppression(unit.content, unit.lineInfo);
+      final isSuppressed = suppression.isSuppressed(_ignoreName);
+      if (isSuppressed) {
+        return {};
+      }
+
+      final visitor = PublicCodeVisitor(suppression, _ignoreName);
       unit.unit.visitChildren(visitor);
 
       return visitor.topLevelElements;