+++ test/resources/unnecessary_nullable_analyzer/suppressed_file.dart
@@ -1,9 +1,8 @@
-// ignore_for_file: avoid-unused-parameters, no-empty-block
+// ignore_for_file: unnecessary-nullable
 
-class ClassWithMethods {
+class IgnoredClassWithMethods {
   void someMethod(String? value) {}
 
-  // LINT
   void alwaysNonNullable(String? anotherValue) {}
 
   void multipleParametersUsed(
@@ -12,14 +11,10 @@ class ClassWithMethods {
     required String? name,
   }) {}
 
-  // LINT
   void multipleParametersWithNamed(
     String? value,
     int anotherValue, {
     required String? name,
     String? secondName,
   }) {}
-
-  // ignore: unnecessary-nullable
-  void ignoredAlwaysNonNullable(String? anotherValue) {}
 }