+++ test/resources/unused_files_analyzer/suppressed_file.dart
@@ -1,0 +1,3 @@
+// ignore_for_file: unused-files
+
+class SomeClass {}